6.5

+++ frontend/orders_list_app/src/views/Orders.vue
@@ -10,47 +10,22 @@
       <template v-slot:top>
         <v-toolbar flat color="white">
           <v-spacer></v-spacer>
-          <CreateOrder v-bind:showSnackBar="showSnackBar" />
+          <CreateOrder :showSnackBar="showSnackBar" />
         </v-toolbar>
       </template>
 
       <template v-slot:expanded-item="{ item }">
         <td :colspan="headers.length">
-          <v-container fluid>
-            <v-row v-for="orderItem in item.orderItems" :key="orderItem.id">
-              <v-col cols="12" sm="6" md="2">
-                <v-img :src="orderItem.photoUrl" max-height="120px"></v-img>
-                <v-divider class="mx-4" inset vertical></v-divider>
-              </v-col>
-              <v-col cols="12" sm="6" md="2">
-                <p>Item.Name : {{orderItem.name}}</p>
-              </v-col>
-              <v-col cols="12" sm="6" md="2">
-                <p>Item.Id : {{orderItem.productId}}</p>
-              </v-col>
-              <v-col cols="12" sm="6" md="2">
-                <p>Item.Price : {{orderItem.productPrice}}</p>
-              </v-col>
-              <v-col cols="12" sm="6" md="2">
-                Item.Quantity :
-                <v-text-field
-                  v-if="editingItemId == item.id"
-                  v-model="orderItem.quantity"
-                  hide-details
-                  single-line
-                  type="number"
-                />
-                <p v-else>{{orderItem.quantity}}</p>
-              </v-col>
-              <v-col cols="12" sm="6" md="2">
-                <v-icon
-                  v-if="editingItemId == item.id"
-                  small
-                  @click="deleteOrderItem(item.id, orderItem.id)"
-                >mdi-delete</v-icon>
-              </v-col>
-            </v-row>
-          </v-container>
+          <OrderItem
+            :item="item"
+            :editingItemId="editingItemId"
+            :deleteOrderItem="deleteOrderItem"
+          />
+          <CreateOrderItem
+            v-if="editingItemId == item.id"
+            :showSnackBar="showSnackBar"
+            :item="item"
+          />
         </td>
       </template>
 
@@ -69,16 +44,28 @@
       </template>
 
       <template v-slot:item.actions="{ item }">
-        <div v-if="editingItemId == item.id">
-          <v-icon small class="mr-2" @click="saveItem(item)">mdi-content-save</v-icon>
-        </div>
-        <div v-if="editingItemId == item.id">
-          <v-icon small class="mr-2" @click="restoreEditItem(item)">mdi-close-circle</v-icon>
-        </div>
-        <div v-else>
-          <v-icon small class="mr-2" @click="editItem(item)">mdi-pencil</v-icon>
-          <v-icon small @click="deleteItem(item)">mdi-delete</v-icon>
-        </div>
+        <v-icon
+          v-if="editingItemId == item.id"
+          medium
+          class="mr-2"
+          @click="saveItem(item)"
+          dense
+        >mdi-content-save</v-icon>
+        <v-icon
+          v-if="editingItemId == item.id"
+          medium
+          class="mr-2"
+          @click="restoreEditItem(item)"
+          dense
+        >mdi-close-circle</v-icon>
+        <v-icon
+          v-if="editingItemId != item.id"
+          medium
+          class="mr-2"
+          @click="editItem(item)"
+          dense
+        >mdi-pencil</v-icon>
+        <v-icon v-if="editingItemId != item.id" medium @click="deleteItem(item)" dense>mdi-delete</v-icon>
       </template>
     </v-data-table>
 
@@ -88,17 +75,25 @@
         <v-btn color="white" text v-bind="attrs" @click="snackbar = false">Close</v-btn>
       </template>
     </v-snackbar>
+
+    <Confirm ref="confirm" />
   </div>
 </template>
 
 <script>
 import { mapState } from "vuex";
-import CreateOrder from "./CreateOrder";
+import CreateOrder from "../components/CreateOrder";
+import CreateOrderItem from "../components/CreateOrderItem";
+import OrderItem from "../components/OrderItem";
+import Confirm from "../components/DialogConfirm";
 
 export default {
   name: "Orders",
   components: {
-    CreateOrder
+    CreateOrder,
+    CreateOrderItem,
+    OrderItem,
+    Confirm
   },
   computed: {
     ...mapState({
@@ -123,6 +118,15 @@ export default {
       this.editingItemId = item.id;
     },
     saveItem: async function(item) {
+      if (
+        !(await this.$refs.confirm.open(
+          "Save",
+          "Do you want to apply changes?",
+          { color: "warning" }
+        ))
+      )
+        return;
+
       this.tableLoading = true;
 
       let itemStatus = this.$store.getters["statuses/getByName"](item.status);
@@ -163,6 +167,15 @@ export default {
       this.cachedItem = null;
     },
     deleteItem: async function(item) {
+      if (
+        !(await this.$refs.confirm.open(
+          "Delete",
+          "Do you want to delete this order?",
+          { color: "red" }
+        ))
+      )
+        return;
+
       this.tableLoading = true;
       try {
         await this.$store.dispatch("orders/deleteItem", item);
@@ -205,3 +218,10 @@ export default {
   })
 };
 </script>
+
+
+<style>
+.v-data-table tbody tr.v-data-table__expanded__content {
+  box-shadow: none;
+}
+</style>
